Test: ListLayout 관련 TC 분리
